rethinking ApplicationConfig for tests
