+++ javers-spring/src/test/groovy/org/javers/spring/auditable/aspect/JaversAuditableRepositoryAspectTest.groovy
@@ -2,8 +2,8 @@ package org.javers.spring.auditable.aspect
 
 import org.aspectj.lang.ProceedingJoinPoint
 import org.javers.common.collections.Lists
-import org.javers.spring.integration.DummyObject
-import org.javers.spring.integration.DummyAuditedCrudRepository
+import org.javers.spring.repository.DummyObject
+import org.javers.spring.repository.DummyAuditedCrudRepository
 import spock.lang.Specification
 
 /**
