+++ javers-spring/src/test/groovy/org/javers/spring/auditable/integration/JaversAuditableAspectIntegrationTest.groovy
@@ -1,9 +1,9 @@
 package org.javers.spring.auditable.integration
 
 import org.javers.core.Javers
-import org.javers.spring.integration.DummyAuditedRepository
-import org.javers.spring.integration.DummyObject
-import org.javers.spring.integration.JaversSpringIntegrationTestApplicationConfig
+import org.javers.spring.example.JaversSpringJpaApplicationConfig
+import org.javers.spring.repository.DummyAuditedRepository
+import org.javers.spring.repository.DummyObject
 import org.springframework.context.ApplicationContext
 import org.springframework.context.annotation.AnnotationConfigApplicationContext
 import spock.lang.Shared
@@ -26,7 +26,7 @@ class JaversAuditableAspectIntegrationTest extends Specification {
     DummyAuditedRepository repository
 
     def setupSpec() {
-        context = new AnnotationConfigApplicationContext(JaversSpringIntegrationTestApplicationConfig)
+        context = new AnnotationConfigApplicationContext(JaversSpringJpaApplicationConfig)
         javers = context.getBean(Javers)
         repository = context.getBean(DummyAuditedRepository)
     }
