+++ javers-spring/src/test/groovy/org/javers/spring/repository/DummyNoAuditCrudRepository.groovy
@@ -1,4 +1,4 @@
-package org.javers.spring.integration
+package org.javers.spring.repository
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
